test: update tests + trait test
