+++ plugins/BEdita/Core/tests/TestCase/Model/Behavior/CategoriesBehaviorTest.php
@@ -56,6 +56,7 @@ class CategoriesBehaviorTest extends TestCase
                         [
                             'name' => 'second-cat',
                             'id' => 2,
+                            'object_type_name' => null,
                         ]
                     ]
                 ],
