+++ plugins/BEdita/Core/tests/TestCase/Model/Entity/PropertyTest.php
@@ -105,88 +105,6 @@ class PropertyTest extends TestCase
     }
 
     /**
-     * Data provider for `testGetObjectTypeName` test case.
-     *
-     * @return array
-     */
-    public function getObjectTypeNameProvider()
-    {
-        return [
-            'document' => [
-                'documents',
-                2,
-            ],
-            'non existent' => [
-                null,
-                -1,
-            ],
-            'invalid' => [
-                null,
-                null,
-            ],
-        ];
-    }
-
-    /**
-     * Test magic getter for object type name property.
-     *
-     * @param string|null $expected Expected object type name.
-     * @param mixed $objectTypeId Object type ID.
-     * @return void
-     *
-     * @covers ::_getObjectTypeName()
-     * @dataProvider getObjectTypeNameProvider()
-     */
-    public function testGetObjectTypeName($expected, $objectTypeId)
-    {
-        $entity = new Property();
-        $entity->object_type_id = $objectTypeId;
-
-        $objectTypeName = $entity->object_type_name;
-
-        static::assertSame($expected, $objectTypeName);
-    }
-
-    /**
-     * Data provider for `testSetObjectTypeName` test case.
-     *
-     * @return array
-     */
-    public function setObjectTypeNameProvider()
-    {
-        return [
-            'document' => [
-                2,
-                'documents',
-            ],
-            'non existent' => [
-                null,
-                'this type does not exist',
-            ],
-        ];
-    }
-
-    /**
-     * Test magic setter for object type name property.
-     *
-     * @param string|null $expected Expected object type ID.
-     * @param mixed $objectTypeName Object type name.
-     * @return void
-     *
-     * @covers ::_setObjectTypeName()
-     * @dataProvider setObjectTypeNameProvider()
-     */
-    public function testSetObjectTypeName($expected, $objectTypeName)
-    {
-        $entity = new Property();
-        $entity->object_type_name = $objectTypeName;
-
-        $objectTypeId = $entity->object_type_id;
-
-        static::assertSame($expected, $objectTypeId);
-    }
-
-    /**
      * Data provider for `testGetPropertyTypeName` test case.
      *
      * @return array
